Cleanups from CR with Lee.
