+++ desktop/cmp/dash/canvas/impl/DashCanvasContextMenu.ts
@@ -21,9 +21,9 @@ import {isEmpty} from 'lodash';
 export const dashCanvasContextMenu = hoistCmp.factory({
     model: null,
     observer: null,
-    render({dashCanvasModel, position, e}) {
+    render({dashCanvasModel, position, contextMenuEvent}) {
         const menuItems = createMenuItems({dashCanvasModel, position});
-        return contextMenu({menuItems, contextMenuEvent: e});
+        return contextMenu({menuItems, contextMenuEvent});
     }
 });
